feat: 새로운 API에 맞춰 DTO 변경
