+++ android/app/src/main/java/com/example/airbnb/ui/search/SearchCloseTravelAdapter.kt
@@ -6,7 +6,6 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.example.airbnb.databinding.ItemCloseTravelBinding
-import com.example.airbnb.network.dto.Region
 
 class SearchCloseTravelAdapter : ListAdapter<Region, SearchCloseTravelAdapter.SearchCloseTravelAdapterHolder>(
     FileDiffCallBack
