+++ android/app/src/main/java/com/example/airbnb/ui/search/ViewModel.kt
@@ -4,7 +4,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.airbnb.data.repository.Repository
 import com.example.airbnb.network.common.NetworkResponse
-import com.example.airbnb.network.dto.Region
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.*
